Restructured files to take interface, ListElement and mapping funcion out. Also added some comments
