+++ src/components/emojiprop.tsx
@@ -1,0 +1,5 @@
+interface emojiProp {
+    emoji: string;
+  }
+
+export default emojiProp
